\name{General Internal Functions}
\alias{.pbdMPIEnv}
\title{ All General Internal Functions }
\description{ All general internal functions }
\references{
  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com}, George Ostrouchov,
  Drew Schmidt, Pragneshkumar Patel, and Hao Yu.
}
\keyword{internal}
