\name{SPMD Internal Functions}
\alias{spmd.allcheck.type}
\alias{spmd.allgather.object}
\alias{spmd.allgather.array}
\alias{spmd.allgather.default}
\alias{spmd.allgatherv.default}
\alias{spmd.allgather.integer}
\alias{spmd.allgather.double}
\alias{spmd.allgather.raw}
\alias{spmd.allgatherv.integer}
\alias{spmd.allgatherv.double}
\alias{spmd.allgatherv.raw}
\alias{spmd.allreduce.object}
\alias{spmd.allreduce.array}
\alias{spmd.allreduce.default}
\alias{spmd.allreduce.integer}
\alias{spmd.allreduce.double}
\alias{spmd.allreduce.logical}
\alias{spmd.allreduce.float}
\alias{spmd.allreduce.float32}
\alias{spmd.bcast.object}
\alias{spmd.bcast.array}
\alias{spmd.bcast.default}
\alias{spmd.bcast.integer}
\alias{spmd.bcast.double}
\alias{spmd.bcast.raw}
\alias{spmd.bcast.string}
\alias{spmd.bcast.message}
\alias{spmd.barrier}
\alias{spmd.comm.set.errhandler}
\alias{spmd.comm.is.null}
\alias{spmd.comm.rank}
\alias{spmd.comm.size}
\alias{spmd.comm.dup}
\alias{spmd.comm.free}
\alias{spmd.init}
\alias{spmd.finalize}
\alias{spmd.is.finalized}
\alias{spmd.is.master}
\alias{spmd.get.processor.name}
\alias{spmd.comm.abort}
\alias{spmd.comm.split}
\alias{spmd.comm.disconnect}
\alias{spmd.comm.connect}
\alias{spmd.comm.accept}
\alias{spmd.port.open}
\alias{spmd.port.close}
\alias{spmd.serv.publish}
\alias{spmd.serv.unpublish}
\alias{spmd.serv.lookup}
\alias{spmd.comm.spawn}
\alias{spmd.comm.get.parent}
\alias{spmd.intercomm.merge}
\alias{spmd.intercomm.create}
\alias{spmd.comm.c2f}
\alias{spmd.gather.object}
\alias{spmd.gather.array}
\alias{spmd.gather.default}
\alias{spmd.gatherv.default}
\alias{spmd.gather.integer}
\alias{spmd.gather.double}
\alias{spmd.gather.raw}
\alias{spmd.gatherv.integer}
\alias{spmd.gatherv.double}
\alias{spmd.gatherv.raw}
\alias{spmd.info.create}
\alias{spmd.info.set}
\alias{spmd.info.free}
\alias{spmd.info.c2f}
\alias{spmd.irecv.default}
\alias{spmd.irecv.integer}
\alias{spmd.irecv.double}
\alias{spmd.irecv.raw}
\alias{spmd.isend.default}
\alias{spmd.isend.integer}
\alias{spmd.isend.double}
\alias{spmd.isend.raw}
\alias{spmd.recv.default}
\alias{spmd.recv.integer}
\alias{spmd.recv.double}
\alias{spmd.recv.raw}
\alias{spmd.reduce.object}
\alias{spmd.reduce.array}
\alias{spmd.reduce.default}
\alias{spmd.reduce.integer}
\alias{spmd.reduce.double}
\alias{spmd.scatter.object}
\alias{spmd.reduce.logical}
\alias{spmd.reduce.float}
\alias{spmd.reduce.float32}
\alias{spmd.scatter.array}
\alias{spmd.scatter.default}
\alias{spmd.scatterv.default}
\alias{spmd.scatter.integer}
\alias{spmd.scatter.double}
\alias{spmd.scatter.raw}
\alias{spmd.scatterv.integer}
\alias{spmd.scatterv.double}
\alias{spmd.scatterv.raw}
\alias{spmd.send.default}
\alias{spmd.send.integer}
\alias{spmd.send.double}
\alias{spmd.send.raw}
\alias{spmd.sendrecv.default}
\alias{spmd.sendrecv.integer}
\alias{spmd.sendrecv.double}
\alias{spmd.sendrecv.raw}
\alias{spmd.sendrecv.replace.default}
\alias{spmd.sendrecv.replace.integer}
\alias{spmd.sendrecv.replace.double}
\alias{spmd.sendrecv.replace.raw}
\alias{spmd.hostinfo}
\alias{spmd.comm.print}
\alias{spmd.comm.cat}
\alias{spmd.probe}
\alias{spmd.iprobe}
\alias{spmd.anysource}
\alias{spmd.anytag}
\alias{spmd.get.sourcetag}
\alias{spmd.get.count}
\alias{spmd.is.comm.null}
\alias{spmd.wait}
\alias{spmd.waitany}
\alias{spmd.waitsome}
\alias{spmd.waitall}
\title{ All SPMD Internal Functions }
\description{ All SPMD internal functions }
\references{
  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com}, George Ostrouchov,
  Drew Schmidt, Pragneshkumar Patel, and Hao Yu.
}
\keyword{internal}
