\name{Comm Internal Functions}
\alias{comm.sort.integer}
\alias{comm.sort.double}
\alias{comm.sort.default}
\alias{comm.dist.common}
\alias{comm.dist.gbd}
\alias{comm.pairwise.common}
\alias{comm.pairwise.gbd}
\title{ All Comm Internal Functions }
\description{ All comm internal functions }
\references{
  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com}, George Ostrouchov,
  Drew Schmidt, Pragneshkumar Patel, and Hao Yu.
}
\keyword{internal}
