\name{Internal Functions}
\alias{g.sort.integer}
\alias{g.sort.double}
\alias{g.sort.default}
\title{ All Internal Functions }
\description{ All internal functions }
\references{
  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com}, George Ostrouchov,
  Drew Schmidt, Pragneshkumar Patel, and Hao Yu.
}
\keyword{internal}

