\name{SPMD Control Functions}
\alias{SPMD.CT}
\alias{SPMD.OP}
\alias{SPMD.IO}
\alias{SPMD.TP}
\alias{.mpiopt_init}

\title{Sets of controls in pbdMPI.}
\description{
  These sets of controls are used to provide default values in this package.
}
\seealso{
  \code{\link{.pbd_env}}.
}
\references{
  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com}, George Ostrouchov,
  Drew Schmidt, Pragneshkumar Patel, and Hao Yu.
}
\keyword{global variables}
